# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Golang admin",
        "title": "go vben admin API",
        "termsOfService": "https://github.com/liluhao/admin",
        "contact": {
            "name": "liluhao",
            "url": "http://foolartist.top",
            "email": "2415306912@qq.com"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:80",
    "paths": {
        "/app/admin/dept/add": {
            "post": {
                "description": "添加部门",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "部门管理"
                ],
                "summary": "添加部门",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "部门信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SysDeptInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/app/admin/dept/list": {
            "get": {
                "description": "获取部门列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "部门管理"
                ],
                "summary": "部门列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "部门名",
                        "name": "deptName",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "状态",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "当前页",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "每页条数",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/dto.SysDeptOutput"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "items": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/dto.SysDept"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.SysDept": {
            "type": "object",
            "required": [
                "deptName"
            ],
            "properties": {
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.SysDept"
                    }
                },
                "createTime": {
                    "type": "string"
                },
                "deptName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "orderNo": {
                    "type": "integer"
                },
                "parentDept": {
                    "type": "string"
                },
                "pid": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "dto.SysDeptInput": {
            "type": "object",
            "required": [
                "deptName"
            ],
            "properties": {
                "deptName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "orderNo": {
                    "type": "integer"
                },
                "parentDept": {
                    "type": "string"
                },
                "pid": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "dto.SysDeptOutput": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.SysDept"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "middleware.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "msg": {
                    "type": "string"
                },
                "result": {
                    "type": "object"
                }
            }
        }
    }
}